requirements.txt
pytest
pytest-html
allure-pytest
PyYAML
selenium
zmail
